Master/dev-requirements.txt
-r requirements.txt
flake8==3.5.0
idna==2.6
urllib3==1.22
chardet==3.0.4
mccabe==0.6.1
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.2.3
requests==2.18.4
simplejson==3.11.1
python-decouple==3.1